Use decode_deterministic_cache_info on Family 15h and later AMD CPUs
